liberação de tiros corrigidas, arrumar lógica dos inimigos, liberar toda a memória, e adicionar inimigo extra

+++ hud.c
@@ -16,7 +16,7 @@ hud* create_hud(){
     return new_hud;
 }
 
-void generate_hud(hud* hud, player* player, int round, ALLEGRO_BITMAP* sprite_sheet, ALLEGRO_FONT* font){
+void generate_hud(hud* hud, player* player, unsigned short round, ALLEGRO_BITMAP* sprite_sheet, ALLEGRO_FONT* font){
    unsigned short lost_lives = player->max_lives - player->lives;
    int i = 0;
    // full hearts
@@ -43,7 +43,7 @@ void generate_hud(hud* hud, player* player, int round, ALLEGRO_BITMAP* sprite_sh
     al_draw_textf(font, al_map_rgb(255, 255, 255), 640 - 100, 20, 0, "%d", player->score);
     // round
     al_draw_text(font, al_map_rgb(255, 255, 255), 640/2 - 100, 0, 0, "ROUND:");
-    al_draw_textf(font, al_map_rgb(255, 255, 255), 640/2 - 100, 20, 0, "%d", round/8);
+    al_draw_textf(font, al_map_rgb(255, 255, 255), 640/2 - 100, 20, 0, "%u", round/8);
 
     // linha de divisão
     al_draw_line(0, 70, 640, 70, al_map_rgb(255, 255, 255), 5);
